feat: resource deletion
Former-commit-id: ea025cc57986a4820ab9e6a1b2ed5c6c3f41286c

+++ quickstart/scripts/cloud-setup/Delete-AzureResources.ps1
@@ -26,10 +26,13 @@ $answer = read-host -prompt "`nPress 'y' to delete all the resources listed abov
 $yesList = 'yes','y'
 
 if ($yesList -contains $answer.ToLower()) {
-    Get-AzResourceGroup | Where-Object ResourceGroupName -match $filter | Remove-AzResourceGroup -AsJob -Force
-    Get-AzADServicePrincipal -DisplayName ("SP-"+$projectName+"-DevTest") | ForEach-Object { Remove-AzADServicePrincipal -ApplicationId $_.ApplicationId -Force }
-
+    if ($myServicePrincipals.Count -gt 0){
+        Get-AzADServicePrincipal -DisplayName ("SP-"+$projectName+"-DevTest") | ForEach-Object { Remove-AzADServicePrincipal -ApplicationId $_.ApplicationId -Force }
+        Get-AzADApplication -DisplayName ("SP-"+$projectName+"-DevTest") | ForEach-Object { Remove-AzADApplication -ApplicationId $_.ApplicationId -Force }
+    }
+    if ($myResourceGroups.Count -gt 0){
+        Get-AzResourceGroup | Where-Object ResourceGroupName -match $filter | Remove-AzResourceGroup -AsJob -Force
+    }
 } else {
     Write-Host "[Command Skipped] Your resources were not deleted."
 }
-
